refactor(network-policy): scaffold package and wire createProxyRuntime
Stage 0 (0a+0b): package skeleton and composition root facade.

- Scaffold @packages/network-policy with port stubs, NetworkExchange, and protocol types
- Add NetworkInterceptionRuntime interface and createProxyRuntime composition root
- Wire ServerBase.createNetworkProxy to createProxyRuntime
