+++ packages/net-stubbing/lib/server/types.ts
@@ -1,24 +1,11 @@
-import type {
-  RouteMatcherOptions,
-  BackendStaticResponse,
-} from '../types'
+import type { BackendRoute } from '@packages/network-policy'
 import type {
   InterceptedRequest,
 } from './intercepted-request'
 
+export type { BackendRoute, GetFixtureFn } from '@packages/network-policy'
+
 export { ResourceType } from '../types'
-
-export type GetFixtureFn = (path: string, opts?: { encoding?: string | null }) => Promise<any>
-
-export interface BackendRoute {
-  routeMatcher: RouteMatcherOptions
-  id: string
-  hasInterceptor: boolean
-  staticResponse?: BackendStaticResponse
-  getFixture: GetFixtureFn
-  matches: number
-  disabled?: boolean
-}
 
 export interface NetStubbingState {
   pendingEventHandlers: {
