+++ packages/network-policy/.eslintignore
@@ -1,0 +1,4 @@
+**/dist
+**/*.d.ts
+**/package-lock.json
+**/tsconfig.json
