+++ packages/network-policy/lib/exchange/network-exchange.ts
@@ -1,0 +1,9 @@
+/**
+ * Normalized request/response snapshot passed from adapters to NetworkPolicyCore.
+ * Fields grow in Stage 3a/3b.
+ */
+export interface NetworkExchange {
+  requestId?: string
+  url?: string
+  method?: string
+}
