+++ packages/network-policy/lib/ports/index.ts
@@ -1,0 +1,3 @@
+export * from './driving-ports'
+
+export * from './driven-ports'
