+++ packages/network-policy/package.json
@@ -1,8 +1,8 @@
 {
-  "name": "@packages/net-stubbing",
+  "name": "@packages/network-policy",
   "version": "0.0.0-development",
   "private": true,
-  "main": "./lib/server",
+  "main": "./lib/index",
   "scripts": {
     "build-prod": "tsc --project .",
     "check-ts": "tsc --noEmit && yarn -s tslint",
@@ -14,22 +14,9 @@
     "tslint": "tslint --config ../ts/tslint.json --project ."
   },
   "dependencies": {
-    "@packages/network-policy": "0.0.0-development",
-    "@packages/socket": "0.0.0-development",
-    "debug": "^4.4.0",
-    "is-html": "^2.0.0",
-    "istextorbinary": "6.0.0",
-    "lodash": "^4.17.15",
-    "mime": "^3.0.0",
-    "minimatch": "^3.1.2",
-    "throttle": "^1.0.3"
+    "@packages/types": "0.0.0-development"
   },
   "devDependencies": {
-    "@packages/errors": "0.0.0-development",
-    "@packages/network": "0.0.0-development",
-    "@packages/proxy": "0.0.0-development",
-    "@packages/telemetry": "0.0.0-development",
-    "@types/mime": "^3.0.1",
     "vitest": "^3.2.4"
   },
   "files": [
