+++ packages/network-policy/tsconfig.json
@@ -1,0 +1,12 @@
+{
+  "extends": "../ts/tsconfig.json",
+  "include": [
+    "*.ts",
+    "lib/*.ts",
+    "lib/**/*.ts",
+    "test/**/*.ts"
+  ],
+  "files": [
+    "./../ts/index.d.ts"
+  ]
+}
