+++ packages/server/package.json
@@ -141,6 +141,7 @@
     "@packages/https-proxy": "0.0.0-development",
     "@packages/launcher": "0.0.0-development",
     "@packages/net-stubbing": "0.0.0-development",
+    "@packages/network-policy": "0.0.0-development",
     "@packages/network": "0.0.0-development",
     "@packages/network-tools": "0.0.0-development",
     "@packages/proxy": "0.0.0-development",
